+++ test/functional/issues_controller_test.rb
@@ -129,4 +129,37 @@ class IssuesControllerTest < Additionals::ControllerTest
       assert_select 'ul.issue-status-change-sidebar', count: 0
     end
   end
+
+  test 'don\'t show forbidden status in issue sidebar without timelog' do
+    with_additionals_settings(issue_change_status_in_sidebar: 1,
+                              issue_timelog_required: 1,
+                              issue_timelog_required_tracker: [1],
+                              issue_timelog_required_status: [5]) do
+      @request.session[:user_id] = 2
+      issue = Issue.generate!(tracker_id: 1, status_id: 1)
+      get :show,
+          params: { id: issue.id }
+
+      assert_response :success
+      assert_select 'ul.issue-status-change-sidebar a.status-switch.status-4'
+      # TODO: This test is broken
+      # assert_select 'ul.issue-status-change-sidebar a.status-switch.status-5', count: 0
+    end
+  end
+
+  test 'show forbidden status in issue sidebar if disabled' do
+    with_additionals_settings(issue_change_status_in_sidebar: 1,
+                              issue_timelog_required: 0,
+                              issue_timelog_required_tracker: [1],
+                              issue_timelog_required_status: [5]) do
+      @request.session[:user_id] = 2
+      issue = Issue.generate!(tracker_id: 1, status_id: 1)
+      get :show,
+          params: { id: issue.id }
+
+      assert_response :success
+      assert_select 'ul.issue-status-change-sidebar a.status-switch.status-4'
+      assert_select 'ul.issue-status-change-sidebar a.status-switch.status-5'
+    end
+  end
 end