+++ app/helpers/additionals_menu_helper.rb
@@ -109,7 +109,7 @@ module AdditionalsMenuHelper
       begin
         plugin_item_base = plugin.id.to_s.camelize.constantize
       rescue LoadError
-        Rails.logger.info 'Ignore plugin for help integration'
+        Rails.logger.debug "Ignore plugin #{plugin.id} for help integration"
       rescue StandardError => e
         raise e unless e.class.to_s == 'NameError'
       end