+++ init.rb
@@ -4,7 +4,7 @@ Redmine::Plugin.register :additionals do
   name 'Additionals'
   author 'AlphaNodes GmbH'
   description 'Customizing Redmine, providing wiki macros and act as a library/function provider for other Redmine plugins'
-  version '2.0.14'
+  version '2.0.15'
   author_url 'https://alphanodes.com/'
   url 'https://github.com/alphanodes/additionals'
 
@@ -24,6 +24,7 @@ Redmine::Plugin.register :additionals do
     permission :edit_closed_issues, {}
     permission :edit_issue_author, {}
     permission :change_new_issue_author, {}
+    permission :issue_timelog_never_required, {}
   end
 
   project_module :time_tracking do