+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - FontAwesome 5.4.1 support
 - Usability improvment for change author in issue formular
 - options and permission for issue requires timelog to use status
+- New option to freeze issues on close. With permission "edit closed issue" user can break this rule.
 
 2.0.14
 ++++++