+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+2.0.15
+++++++
+
+- FontAwesome 5.5.0 support
+- Usability improvement for change author in issue formular
+- options and permission for issue requires timelog to use status
+- New option to freeze issues on close. With permission "edit closed issue" user can break this rule.
+- Fix problem with help menu and other redmine plugins (compatibility problem with other plugins)
+- Fix problem with disabled users and changing author for issues
+
 2.0.14
 ++++++